fix: SettingsProfile removed; enums moved to the same catalog; SettingsNamesEnum added
