+++ src/main/java/ru/aston/gamerent/model/entity/Role.java
@@ -12,6 +12,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
+import ru.aston.gamerent.model.entity.enums.RoleNameEnum;
 
 import java.util.Objects;
