+++ src/main/java/ru/aston/gamerent/model/entity/Setting.java
@@ -1,10 +1,21 @@
 package ru.aston.gamerent.model.entity;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
+import ru.aston.gamerent.model.entity.enums.SettingsNamesEnum;
 
 import java.util.List;
 import java.util.Objects;
@@ -12,7 +23,9 @@ import java.util.Objects;
 @Entity
 @Getter
 @Setter
+@Builder
 @NoArgsConstructor
+@AllArgsConstructor
 @ToString
 @Table(name = "settings")
 public class Setting {
@@ -21,7 +34,8 @@ public class Setting {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @Column(name = "setting_name", unique = true, nullable = false, columnDefinition = "CHARACTER VARYING(30)")
-    private String settingName;
+    @Enumerated(EnumType.STRING)
+    private SettingsNamesEnum settingName;
     @OneToMany(mappedBy = "setting")
     @ToString.Exclude
     private List<SettingValue> settingsProfiles;
