+++ src/main/java/ru/aston/gamerent/model/entity/enums/RoleNameEnum.java
@@ -1,4 +1,4 @@
-package ru.aston.gamerent.model.entity;
+package ru.aston.gamerent.model.entity.enums;
 
 public enum RoleNameEnum {
     ROLE_ADMIN,
