+++ src/main/java/ru/aston/gamerent/model/entity/enums/SettingsNamesEnum.java
@@ -1,0 +1,9 @@
+package ru.aston.gamerent.model.entity.enums;
+
+public enum SettingsNamesEnum {
+
+    TIME_ENDING_NOTIFICATIONS,
+    SALES_NOTIFICATIONS,
+    NEW_ARRIVALS_NOTIFICATIONS,
+    NEWS
+}
