+++ src/main/resources/db.changelog/changelog.yml
@@ -1,3 +1,5 @@
 databaseChangeLog:
   - include:
       file: db.changelog/changelogs/GAMERENT-1-25-06-2023-changelog.yaml
+  - include:
+      file: db.changelog/changelogs/GAMERENT-2-26-06-2023-changelog.yaml
