+++ src/main/resources/db.changelog/changelogs/GAMERENT-1-25-06-2023-changelog.yaml
@@ -109,105 +109,3 @@ databaseChangeLog:
               onUpdate: CASCADE
               referencedColumnNames: id
               referencedTableName: roles
-  - changeSet:
-      id: 4
-      author: yuri
-      changes:
-        - createTable:
-            tableName: settings_profiles
-            columns:
-              - column:
-                  name: id
-                  type: bigint
-                  autoIncrement: true
-                  constraints:
-                    nullable: false
-                    primaryKey: true
-                    primaryKeyName: pk_settings_profiles
-  - changeSet:
-      id: 5
-      author: yuri
-      changes:
-        - createTable:
-            tableName: settings
-            columns:
-              - column:
-                  name: id
-                  type: bigint
-                  autoIncrement: true
-                  constraints:
-                    nullable: false
-                    primaryKey: true
-                    primaryKeyName: pk_settings
-              - column:
-                  name: setting_name
-                  type: varchar(30)
-                  constraints:
-                    unique: true
-                    nullable: false
-  - changeSet:
-      id: 6
-      author: yuri
-      changes:
-        - createTable:
-            tableName: settings_values
-            columns:
-              - column:
-                  name: id
-                  type: bigint
-                  autoIncrement: true
-                  constraints:
-                    nullable: false
-                    primaryKey: true
-                    primaryKeyName: pk_setting_values
-              - column:
-                  name: profile_id
-                  type: bigint
-                  constraints:
-                    nullable: false
-              - column:
-                  name: setting_id
-                  type: bigint
-                  constraints:
-                    nullable: false
-              - column:
-                  name: value
-                  type: boolean
-                  constraints:
-                    nullable: false
-        - addForeignKeyConstraint:
-            baseColumnNames: setting_id
-            baseTableName: settings_values
-            constraintName: fk_vlg2stg
-            onDelete: CASCADE
-            onUpdate: CASCADE
-            referencedColumnNames: id
-            referencedTableName: settings
-        - addForeignKeyConstraint:
-            baseColumnNames: profile_id
-            baseTableName: settings_values
-            constraintName: fk_vlg2pfl
-            onDelete: CASCADE
-            onUpdate: CASCADE
-            referencedColumnNames: id
-            referencedTableName: settings_profiles
-  - changeSet:
-      id: 7
-      author: yuri
-      changes:
-        - addColumn:
-            tableName: users
-            columns:
-              - column:
-                  name: profile_id
-                  type: bigint
-                  constraints:
-                    nullable: false
-        - addForeignKeyConstraint:
-            baseColumnNames: profile_id
-            baseTableName: users
-            constraintName: fk_usr2pfl
-            onDelete: CASCADE
-            onUpdate: CASCADE
-            referencedColumnNames: id
-            referencedTableName: settings_profiles
