+++ src/main/resources/db.changelog/changelogs/GAMERENT-2-26-06-2023-changelog.yaml
@@ -1,0 +1,68 @@
+databaseChangeLog:
+- changeSet:
+    id: 1
+    author: y.kaleganov@astondevs.ru
+    changes:
+    - createTable:
+        tableName: settings
+        columns:
+        - column:
+            name: id
+            type: bigint
+            autoIncrement: true
+            constraints:
+              nullable: false
+              primaryKey: true
+              primaryKeyName: pk_settings
+        - column:
+            name: setting_name
+            type: varchar(30)
+            constraints:
+              unique: true
+              nullable: false
+- changeSet:
+    id: 2
+    author: y.kaleganov@astondevs.ru
+    changes:
+    - createTable:
+        tableName: settings_values
+        columns:
+        - column:
+            name: id
+            type: bigint
+            autoIncrement: true
+            constraints:
+              nullable: false
+              primaryKey: true
+              primaryKeyName: pk_setting_values
+        - column:
+            name: user_id
+            type: bigint
+            constraints:
+              nullable: false
+        - column:
+            name: setting_id
+            type: bigint
+            constraints:
+              nullable: false
+        - column:
+            name: value
+            type: boolean
+            constraints:
+              nullable: false
+    - addForeignKeyConstraint:
+        baseColumnNames: setting_id
+        baseTableName: settings_values
+        constraintName: fk_val2stg
+        onDelete: CASCADE
+        onUpdate: CASCADE
+        referencedColumnNames: id
+        referencedTableName: settings
+    - addForeignKeyConstraint:
+        baseColumnNames: user_id
+        baseTableName: settings_values
+        constraintName: fk_vlg2pfl
+        onDelete: CASCADE
+        onUpdate: CASCADE
+        referencedColumnNames: id
+        referencedTableName: users
